+++ Cargo.toml
@@ -8,9 +8,9 @@ axum = "0.8"
 base64 = "0.22"
 chrono = "0.4"
 clap = { version = "4.6", features = ["derive", "env"] }
-defguard_certs = { git = "https://github.com/DefGuard/defguard.git", rev = "dc0b70e69335c4be98d5b17615003922c9464994" }
-defguard_grpc_tls = { git = "https://github.com/DefGuard/defguard.git", rev = "dc0b70e69335c4be98d5b17615003922c9464994" }
-defguard_version = { git = "https://github.com/DefGuard/defguard.git", rev = "dc0b70e69335c4be98d5b17615003922c9464994" }
+defguard_certs = { git = "https://github.com/DefGuard/defguard.git", rev = "f5ae25dc6810b92ce9bd8dafddefc03f336a69bf" }
+defguard_grpc_tls = { git = "https://github.com/DefGuard/defguard.git", rev = "f5ae25dc6810b92ce9bd8dafddefc03f336a69bf" }
+defguard_version = { git = "https://github.com/DefGuard/defguard.git", rev = "f5ae25dc6810b92ce9bd8dafddefc03f336a69bf" }
 defguard_wireguard_rs = "0.10"
 rustls = { version = "0.23", default-features = false, features = ["ring"] }
 rustls-pki-types = "1"
